refactor: Move diff and logWithInstance to libs/utils
